+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - Add report issue button in settings
+- Fullscreen hides top bar and tabs
 
 ## v0.14.0 - 2026-03-22